Verify creation of dialogporten dialog before publishing correspondence (#784)
* Added a validation for DialogId externalReference on the correspondence before publishing

* Added unit-test that verifies publishing a correspondence fails if its missing dialogPorten reference

* Fixed coderabbit nitpick problem

* Temporary change to manually test publish failure and retry when dialog has not been created

* Added logic to ensure PublishCorrespondence runs after CreateDialogPortenDialog has completed

* Small fix for test

* Attempting to ensure publish happens after Dialog creation by scheduling publish after create dialog job finish

* Schedule in method to fix exception

* Removed delayed DialogPortenDialog creeation that was used for testing

* Avoid attempting to purge dialog for correspondence if the dialog for the correspondence failed to create

* Updated tests to reflect dialog not being purged when dialog was not created for the correspondence on publishfail

---------

+++ src/Altinn.Correspondence.Application/PublishCorrespondence/PublishCorrespondenceHandler.cs
@@ -35,6 +35,7 @@ public class PublishCorrespondenceHandler(
         {
             return AuthorizationErrors.CouldNotFindPartyUuid;
         }
+        bool hasDialogportenDialog = correspondence.ExternalReferences.Any(reference => reference.ReferenceType == ReferenceType.DialogportenDialogId);
         var errorMessage = "";
         if (correspondence == null)
         {
@@ -69,6 +70,10 @@ public class PublishCorrespondenceHandler(
         {
             errorMessage = $"Correspondence {correspondenceId} not visible yet";
         }
+        else if (!hasDialogportenDialog)
+        {
+            errorMessage = $"Dialogporten dialog not created for correspondence {correspondenceId}";
+        }
         else if (correspondence.IgnoreReservation != true && correspondence.GetRecipientUrn().IsSocialSecurityNumber())
         {
             var isReserved = await contactReservationRegistryService.IsPersonReserved(correspondence.GetRecipientUrn().WithoutPrefix());
@@ -98,7 +103,10 @@ public class PublishCorrespondenceHandler(
                 {
                     backgroundJobClient.Enqueue<CancelNotificationHandler>(handler => handler.Process(null, correspondenceId, null, cancellationToken));
                 }
-                backgroundJobClient.Enqueue<IDialogportenService>(dialogportenService => dialogportenService.PurgeCorrespondenceDialog(correspondenceId));
+                if (hasDialogportenDialog)
+                {
+                    backgroundJobClient.Enqueue<IDialogportenService>(dialogportenService => dialogportenService.PurgeCorrespondenceDialog(correspondenceId));
+                }
             }
             else
             {
